+++ app/src/main/java/org/simple/clinic/patient/OngoingNewPatientEntry.kt
@@ -51,7 +51,7 @@ data class OngoingNewPatientEntry(
     val address: Address? = null,
     val phoneNumber: PhoneNumber? = null,
     val identifier: Identifier? = null,
-    val alternativeId: Identifier? = null,
+    val alternateId: Identifier? = null,
     val reminderConsent: ReminderConsent = Granted
 ) : Parcelable {
 
@@ -111,7 +111,7 @@ data class OngoingNewPatientEntry(
       copy(reminderConsent = reminderConsent)
 
   fun withAlternativeId(identifier: Identifier): OngoingNewPatientEntry =
-      copy(alternativeId = identifier)
+      copy(alternateId = identifier)
 
   fun withStreetAddress(streetAddress: String): OngoingNewPatientEntry =
       copy(address = addressOrBlank().copy(streetAddress = streetAddress))
@@ -119,7 +119,11 @@ data class OngoingNewPatientEntry(
   fun withZone(zone: String): OngoingNewPatientEntry =
       copy(address = addressOrBlank().copy(zone = zone))
 
-  fun withPatientPrefillInfo(patientProfileInfo: PatientPrefillInfo, identifier: Identifier, dateTimeFormatter: DateTimeFormatter): OngoingNewPatientEntry =
+  fun withPatientPrefillInfo(
+      patientProfileInfo: PatientPrefillInfo,
+      alternateId: Identifier,
+      dateTimeFormatter: DateTimeFormatter
+  ): OngoingNewPatientEntry =
       copy(
           personalDetails = PersonalDetails(
               fullName = patientProfileInfo.fullName,
@@ -127,7 +131,7 @@ data class OngoingNewPatientEntry(
               gender = patientProfileInfo.gender,
               age = null),
           address = addressOrBlank().withColonyOrVillage(patientProfileInfo.address),
-          identifier = identifier)
+          alternateId = alternateId)
 
   fun validationErrors(
       dobValidator: UserInputDateValidator,