+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Bump Kotlin to v1.5.10
 - Bump AGP to v4.2.1
 - Use `.editorconfig` for project code style
+- Prefill NHID as `alternateId` in `PatientEntryScreen`
 
 ## Features
 - Add NHID support in `PatientEntryScreen`